requirements.txt
python-fasthtml
starlette
jinja2
uvicorn
requests
